Update Guess the secret number (Brute Force 256)
In this lottery contract, we have to guess the secret number. From the code, we can see that the number is represented in the form of a cryptographic hash. If we run this brute force code in remix and then deploy it, we can reveal the hidden number by using the test function. All we have to do the contract and put the number in.