Create Guess the random number
This task took a long time to figure out, but was very easy in the end. From the contract code, we can see that the answer is a public global variable -

answer = uint8(keccak256(block.blockhash(block.number - 1), now))

therefore, if we click on the contract address/creator txn hash/state we can see the differences in the states before and after. By clicking on the dropdown in the 'after' box, we can change this hex to 'number 'and reveal the number we need. All we have to do then is deploy the contract with the correct address and 1 eth to complete the challenge.